backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image processing
opencv-python==4.9.0.80
scikit-image==0.22.0
Pillow==10.2.0

# Machine learning
torch>=2.2.0
torchvision>=0.17.0
scikit-learn==1.4.0
umap-learn==0.5.5

# Graph analysis
networkx==3.2.1

# Data handling
numpy==1.26.3
pandas==2.1.4
scipy==1.12.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0

# Gemini API
google-generativeai==0.4.0
aiohttp==3.9.3
httpx==0.28.1

